tweak bottom spacing to match to spacing
First child is typically an `h2` with a font-size of 1.75rem, so the spacing at the top of the content block was slightly bigger than the spacing at the bottom. The fix is to add a bit of margin on the last child (typically a paragraph).

To make similar computations simpler in the future, I've added a couple of CSS variables: `--lsp` for the line spacing (= line height - 1 = 0.5), and `--hlsp` for half of the line spacing (= 0.25). The extra spacing needed on element A to match the spacing of element B can now be calculated as follows: `calc(var(--hlsp) * (<font-size-of-B> - <font-size-of-A>))`